+++ src/ResourceManager/ServiceBus/ChangeLog.md
@@ -19,6 +19,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix issue with Default Resource Group in CloudShell
 
 ## Version 0.6.2
 * Added EnableBatchedOperations property to Queue